Move definition of row level security to end of model definition
pgModeler can't keep track of dependencies declared in custom SQL.
Therefore the generated SQL can turn out incorrect if custom SQL
code refers to object not yet created. This turned out to be the
case with row level security definitions.
An options is to spend a day switching object ID's, but there is
no guarantee that future changes/updates don't create the same
problem. As such, I decided to move it to the end of the definition.